backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
pytesseract==0.3.10
transformers==4.35.0
torch==2.1.0
torchvision==0.16.0
opencv-python==4.8.1.78
numpy==1.26.2
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
aiofiles==23.2.1
easyocr==1.7.1

# PDF Processing Libraries
pdf2image==1.16.3
PyMuPDF==1.23.8